Adds support for a `statusCodes` configuration option.
This allows for the options object to accept other settings without the assuming every key at the root of the options object is for an HTTP status code. It moves the default configuration for status codes to the `statusCodes` option and maintains backwards compatibility by continuing to check for status code configuration at the root of the options object, but only if it doesn’t exist under `statusCodes`.